[fe#15] refactor: 기획전 컴포넌트 분리
